Cleanup persistence test setup.
