+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/persistence/AbstractDatabasePersistenceWebFailoverTestCase.java
@@ -26,13 +26,12 @@ import java.net.URL;
 
 import org.infinispan.configuration.cache.CacheMode;
 import org.infinispan.transaction.TransactionMode;
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.container.test.api.TargetsContainer;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.as.test.clustering.ClusterTestUtil;
 import org.jboss.as.test.clustering.cluster.web.AbstractWebFailoverTestCase;
 import org.jboss.as.test.clustering.single.web.Mutable;
 import org.jboss.as.test.clustering.single.web.SimpleServlet;
+import org.jboss.as.test.shared.CLIServerSetupTask;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
@@ -46,33 +45,14 @@ import org.junit.runner.RunWith;
  */
 @RunWith(Arquillian.class)
 public abstract class AbstractDatabasePersistenceWebFailoverTestCase extends AbstractWebFailoverTestCase {
-    private static final String DEPLOYMENT_NAME = AbstractDatabasePersistenceWebFailoverTestCase.class.getSimpleName() + ".war";
 
-    public AbstractDatabasePersistenceWebFailoverTestCase() {
+    public AbstractDatabasePersistenceWebFailoverTestCase(String moduleName) {
         // Use NON_TRANSACTIONAL to workaround for ISPN-10029
-        super(DEPLOYMENT_NAME, CacheMode.INVALIDATION_SYNC, TransactionMode.TRANSACTIONAL);
+        super(moduleName, CacheMode.INVALIDATION_SYNC, TransactionMode.TRANSACTIONAL);
     }
 
-    @Deployment(name = DEPLOYMENT_1, managed = false, testable = false)
-    @TargetsContainer(NODE_1)
-    public static Archive<?> deployment1() {
-        return getDeployment();
-    }
-
-    @Deployment(name = DEPLOYMENT_2, managed = false, testable = false)
-    @TargetsContainer(NODE_2)
-    public static Archive<?> deployment2() {
-        return getDeployment();
-    }
-
-    @Deployment(name = DEPLOYMENT_3, managed = false, testable = false)
-    @TargetsContainer(NODE_3)
-    public static Archive<?> deployment3() {
-        return getDeployment();
-    }
-
-    private static Archive<?> getDeployment() {
-        WebArchive war = ShrinkWrap.create(WebArchive.class, DEPLOYMENT_NAME);
+    static Archive<?> getDeployment(String moduleName) {
+        WebArchive war = ShrinkWrap.create(WebArchive.class, moduleName + ".war");
         war.addClasses(SimpleServlet.class, Mutable.class);
         ClusterTestUtil.addTopologyListenerDependencies(war);
         war.setWebXML(AbstractWebFailoverTestCase.class.getPackage(), "web.xml");
@@ -83,5 +63,18 @@ public abstract class AbstractDatabasePersistenceWebFailoverTestCase extends Abs
     @Override
     public void testGracefulSimpleFailover(URL baseURL1, URL baseURL2, URL baseURL3) {
         // TODO rework to use external database process since H2's AUTO_SERVER doesn't handle server restarts reliably
+    }
+
+    public static class ServerSetupTask extends CLIServerSetupTask {
+        public ServerSetupTask() {
+            this.builder.node(THREE_NODES)
+                    .setup("/subsystem=datasources/data-source=web-sessions-ds:add(jndi-name=\"java:jboss/datasources/web-sessions-ds\",enabled=true,use-java-context=true,connection-url=\"jdbc:h2:file:./target/h2/web-sessions;AUTO_SERVER=true;DB_CLOSE_ON_EXIT=FALSE;DB_CLOSE_DELAY=-1;\",driver-name=h2")
+                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence:add")
+                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/store=jdbc:add(data-source=web-sessions-ds,fetch-state=false,purge=false,passivation=false,shared=true)")
+                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/component=transaction:add(mode=BATCH)")
+                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/component=locking:add(isolation=REPEATABLE_READ)")
+                    .teardown("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence:remove")
+                    .teardown("/subsystem=datasources/data-source=web-sessions-ds:remove");
+        }
     }
 }
