+++ testsuite/integration/clustering/src/test/java/org/jboss/as/test/clustering/cluster/web/persistence/CoarseDatabasePersistenceWebFailoverTestCase.java
@@ -22,29 +22,54 @@
 
 package org.jboss.as.test.clustering.cluster.web.persistence;
 
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.container.test.api.TargetsContainer;
 import org.jboss.as.arquillian.api.ServerSetup;
 import org.jboss.as.test.shared.CLIServerSetupTask;
+import org.jboss.shrinkwrap.api.Archive;
 
 /**
  * @author Paul Ferraro
  */
-@ServerSetup(CoarseDatabasePersistenceWebFailoverTestCase.ServerSetupTask.class)
+@ServerSetup({ CoarseDatabasePersistenceWebFailoverTestCase.ServerSetupTask.class, AbstractDatabasePersistenceWebFailoverTestCase.ServerSetupTask.class })
 @org.junit.Ignore("Fails intermittently due to ISPN-10029")
 public class CoarseDatabasePersistenceWebFailoverTestCase extends AbstractDatabasePersistenceWebFailoverTestCase {
+
+    private static final String MODULE_NAME = CoarseDatabasePersistenceWebFailoverTestCase.class.getSimpleName();
+
+    @Deployment(name = DEPLOYMENT_1, managed = false, testable = false)
+    @TargetsContainer(NODE_1)
+    public static Archive<?> deployment1() {
+        return getDeployment();
+    }
+
+    @Deployment(name = DEPLOYMENT_2, managed = false, testable = false)
+    @TargetsContainer(NODE_2)
+    public static Archive<?> deployment2() {
+        return getDeployment();
+    }
+
+    @Deployment(name = DEPLOYMENT_3, managed = false, testable = false)
+    @TargetsContainer(NODE_3)
+    public static Archive<?> deployment3() {
+        return getDeployment();
+    }
+
+    private static Archive<?> getDeployment() {
+        return getDeployment(MODULE_NAME);
+    }
+
+    public CoarseDatabasePersistenceWebFailoverTestCase() {
+        super(MODULE_NAME);
+    }
 
     public static class ServerSetupTask extends CLIServerSetupTask {
         public ServerSetupTask() {
             this.builder.node(THREE_NODES)
-                    .setup("/subsystem=datasources/data-source=web-sessions-ds:add(jndi-name=\"java:jboss/datasources/web-sessions-ds\",enabled=true,use-java-context=true,connection-url=\"jdbc:h2:file:./target/h2/web-sessions;AUTO_SERVER=true;DB_CLOSE_ON_EXIT=FALSE;DB_CLOSE_DELAY=-1;\",driver-name=h2")
-                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence:add")
-                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/store=jdbc:add(data-source=web-sessions-ds,fetch-state=false,purge=false,passivation=false,shared=true")
-                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/component=transaction:add(mode=BATCH)")
-                    .setup("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence/component=locking:add(isolation=REPEATABLE_READ)")
                     .setup("/subsystem=distributable-web/infinispan-session-management=database:add(cache-container=web, cache=database-persistence, granularity=SESSION)")
                     .setup("/subsystem=distributable-web/infinispan-session-management=database/affinity=local:add()")
                     .teardown("/subsystem=distributable-web/infinispan-session-management=database:remove")
-                    .teardown("/subsystem=infinispan/cache-container=web/invalidation-cache=database-persistence:remove")
-                    .teardown("/subsystem=datasources/data-source=web-sessions-ds:remove");
+                    ;
         }
     }
 }
